requirements.txt
# requirements.txt — OpenBrain Python dependencies
openai>=1.0.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=12.0
